moved api base urls to respective files
