+++ src/components/Convert/routes/oneInch.ts
@@ -1,11 +1,19 @@
 import axios from "axios";
 import { ethers } from "ethers";
-import { config } from "../../..";
+import { config, Network, NetworkMap } from "../../..";
 import { BASIS_POINTS_DIVISOR } from "../../../config/hlp";
-import { get1InchBaseUrl } from "../baseApiUrls";
 import { ConvertQuoteInput, ConvertTransactionInput, Quote } from "../Convert";
 import { getApiFeeAsPercentage } from "../getApiFeeAsPercentage";
 import { ONE_INCH_WEIGHT, WeightInput } from "./weights";
+
+export const get1InchBaseUrl = (network: Network) => {
+  const networkNameToIdMap: NetworkMap<number> = {
+    ethereum: 1,
+    polygon: 137,
+    arbitrum: 42161
+  };
+  return `https://api.1inch.exchange/v4.0/${networkNameToIdMap[network]}`;
+};
 
 type OneInchQuoteParams = {
   fromTokenAddress: string;
