+++ src/components/Convert/routes/zeroX.ts
@@ -2,10 +2,13 @@ import axios from "axios";
 import { ConvertQuoteInput, ConvertTransactionInput, Quote } from "../Convert";
 import { getApiFeeAsPercentage } from "../getApiFeeAsPercentage";
 import config from "../../../config";
-import { get0xBaseUrl } from "../baseApiUrls";
 import { BASIS_POINTS_DIVISOR } from "../../../config/hlp";
 import { WeightInput, ZERO_X_WEIGHT } from "./weights";
 import { ethers } from "ethers";
+import { Network } from "../../..";
+
+export const get0xBaseUrl = (network: Network) =>
+  `https://${network === "ethereum" ? "" : network + "."}api.0x.org/swap/v1`;
 
 type ZeroXQuoteParams = {
   buyToken: string;
